Lots of CLI improvements and adding in LLVM integration
