+++ forge-core/include/forge/common/log.h
@@ -19,17 +19,27 @@
 #include <forge/common/types.h>
 #include <stdbool.h>
 
+frg_status_t frg_log_set_minimum_severity(frg_log_severity_t severity);
+
 /// \brief Log a message prefix that indicates that the following message is from the
 ///        source file \a filename.
-frg_status_t frg_log_prefix_source_file(const char* filename);
+frg_status_t frg_log_prefix_source_file(
+    frg_log_severity_t severity,
+    const char* filename
+);
 
 /// \brief Log a message prefix that indicates that the following message is from the
 ///        source file \a filename on line \a lineno.
-frg_status_t frg_log_prefix_source_line(const char* filename, frg_lineno_t lineno);
+frg_status_t frg_log_prefix_source_line(
+    frg_log_severity_t severity,
+    const char* filename,
+    frg_lineno_t lineno
+);
 
 /// \brief Log a message prefix that indicates that the following message is from the
 ///        source file \a filename on line \a lineno at column \a columnno.
 frg_status_t frg_log_prefix_source_char(
+    frg_log_severity_t severity,
     const char* filename,
     frg_lineno_t lineno,
     frg_columnno_t columnno
@@ -37,9 +47,13 @@ frg_status_t frg_log_prefix_source_char(
 
 /// \brief Log a message prefix that indicates that the following message is from the
 ///        current line of the Forge compiler source code.
-#define frg_log_prefix_internal() _frg_log_prefix_internal(__FILE__, __LINE__)
+#define frg_log_prefix_internal(severity) _frg_log_prefix_internal((severity), __FILE__, __LINE__)
 
-void _frg_log_prefix_internal(const char* filename, frg_lineno_t lineno);
+frg_status_t _frg_log_prefix_internal(
+    frg_log_severity_t severity,
+    const char* filename,
+    frg_lineno_t lineno
+);
 
 /// \brief Log a message.
 frg_status_t frg_log(frg_log_severity_t severity, const char* format, ...);
