+++ forge-core/include/forge/config/commands/common.h
@@ -18,14 +18,7 @@
 #include <forge/common/enums.h>
 #include <glib.h>
 
-typedef struct {
-    bool version_short;
-} frg_config_t;
-
-frg_status_t frg_config_new_default(frg_config_t** config);
-void frg_config_destroy(frg_config_t** config);
-
-frg_status_t frg_config_parse_cli(int* exit_status, frg_config_t* config, int argc, const char** argv);
-frg_status_t frg_config_parse_env(frg_config_t* config);
-
-frg_status_t frg_config_log_debug(const frg_config_t* config);
+frg_status_t frg_config_commands_get_single_source_file(
+    const char** path,
+    GList* pos_args
+);
