+++ forge-core/include/forge/config/commands/compile.h
@@ -15,17 +15,6 @@
 
 #pragma once
 
-#include <forge/common/enums.h>
-#include <glib.h>
+#include <forge/cli/command.h>
 
-typedef struct {
-    bool version_short;
-} frg_config_t;
-
-frg_status_t frg_config_new_default(frg_config_t** config);
-void frg_config_destroy(frg_config_t** config);
-
-frg_status_t frg_config_parse_cli(int* exit_status, frg_config_t* config, int argc, const char** argv);
-frg_status_t frg_config_parse_env(frg_config_t* config);
-
-frg_status_t frg_config_log_debug(const frg_config_t* config);
+frg_status_t frg_config_commands_new_compile(frg_cli_command_t** command);
