+++ forge-core/lib/forge/cli/command.c
@@ -16,7 +16,7 @@
 #include <forge/common/color.h>
 #include <forge/common/log.h>
 #include <forge/common/memory.h>
-#include <forge/config/cli/command.h>
+#include <forge/cli/command.h>
 
 frg_status_t frg_cli_command_new(
     frg_cli_command_t** command,
@@ -48,7 +48,7 @@ frg_status_t frg_cli_command_new(
         &(*command)->option_set
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create option set: %s", frg_status_to_string(result)); 
         return result;
     }
@@ -67,7 +67,7 @@ frg_status_t frg_cli_command_destroy(
         &(*command)->option_set
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy option set: %s", frg_status_to_string(result)); 
         return result;
     }
