+++ forge-core/lib/forge/cli/option_set.c
@@ -15,7 +15,7 @@
 
 #include <forge/common/log.h>
 #include <forge/common/memory.h>
-#include <forge/config/cli/option_set.h>
+#include <forge/cli/option_set.h>
 
 frg_status_t _frg_cli_get_short_name_index(int* index, char short_name) {
     if (index == NULL) {
@@ -62,7 +62,7 @@ frg_status_t frg_cli_option_set_destroy(
     for (GList* option = (*option_set)->options; option != NULL; option = option->next) {
         frg_status_t result = frg_cli_option_destroy((frg_cli_option_t**)&option->data);
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy option: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -89,7 +89,7 @@ frg_status_t frg_cli_option_set_add_option(
         int index = 0;
         frg_status_t result = _frg_cli_get_short_name_index(&index, option->short_name);
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to get short name index: %s", frg_status_to_string(result));
             return result;
         }
@@ -146,13 +146,13 @@ frg_status_t frg_cli_option_set_get_option_by_short_name(
     int index = -1;
     frg_status_t result = _frg_cli_get_short_name_index(&index, name);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to get short name index: %s", frg_status_to_string(result)); 
         return result;
     }
 
     if (index < 0 || index >= FRG_CLI_OPTION_SHORT_NAME_MAX_COUNT) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "invalid index %i", index);
         return FRG_STATUS_ERROR_KEY_NOT_FOUND;
     }
@@ -184,7 +184,7 @@ frg_status_t frg_cli_option_set_print_help(
 
         frg_status_t result = frg_cli_option_print_help((const frg_cli_option_t*)option->data);
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print option help: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -223,7 +223,7 @@ frg_status_t _frg_cli_option_set_parse_next_long(
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unknown argument '--%s'", name);
         return FRG_STATUS_CLI_ERROR;
     } else if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to get option by long name: %s", frg_status_to_string(result)); 
         return result;
     }
@@ -261,7 +261,7 @@ frg_status_t _frg_cli_option_set_parse_next_short(
         name
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to get option by short name: %s", frg_status_to_string(result)); 
         return result;
     }
