+++ forge-core/lib/forge/cli/program.c
@@ -16,7 +16,7 @@
 #include <forge/common/color.h>
 #include <forge/common/log.h>
 #include <forge/common/memory.h>
-#include <forge/config/cli/program.h>
+#include <forge/cli/program.h>
 
 frg_status_t frg_cli_program_new(
     frg_cli_program_t** program,
@@ -62,7 +62,7 @@ frg_status_t frg_cli_program_new(
         &(*program)->global_options
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create option set: %s", frg_status_to_string(result)); 
         return result;
     }
@@ -81,7 +81,7 @@ frg_status_t frg_cli_program_destroy(
     for (GList* command = (*program)->commands; command != NULL; command = command->next) {
         frg_status_t result = frg_cli_command_destroy((frg_cli_command_t**)&command->data);
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy command: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -95,7 +95,7 @@ frg_status_t frg_cli_program_destroy(
         &(*program)->global_options
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy option set: %s", frg_status_to_string(result)); 
         return result;
     }
@@ -153,7 +153,7 @@ frg_status_t frg_cli_program_print_help(
 
     frg_status_t result = frg_cli_program_print_version_long(program);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print long version: %s", frg_status_to_string(result)); 
         return result;
     }
@@ -197,7 +197,7 @@ frg_status_t frg_cli_program_print_help(
             program->global_options
         );
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print option set help: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -257,7 +257,7 @@ frg_status_t frg_cli_program_print_help(
             program->global_options
         );
         if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print option set help: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -271,7 +271,7 @@ frg_status_t frg_cli_program_print_help(
 
             result = frg_cli_option_set_print_help(command->option_set);
             if (result != FRG_STATUS_OK) {
-                frg_log_prefix_internal();
+                frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print option set help: %s", frg_status_to_string(result)); 
                 return result;
             }
@@ -350,7 +350,7 @@ frg_status_t frg_cli_program_parse(
             if (result == FRG_STATUS_CLI_ERROR) {
                 return FRG_STATUS_CLI_ERROR;
             } else if (result != FRG_STATUS_OK) {
-                frg_log_prefix_internal();
+                frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
                 frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to parse next argument: %s", frg_status_to_string(result)); 
                 return result;
             }
@@ -386,7 +386,7 @@ frg_status_t frg_cli_program_parse(
             frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unknown command '%s'", argv[argi]);
             return FRG_STATUS_CLI_ERROR;
         } else if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
+            frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
             frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to get command by name: %s", frg_status_to_string(result)); 
             return result;
         }
@@ -409,7 +409,7 @@ frg_status_t frg_cli_program_parse(
                 if (result == FRG_STATUS_CLI_ERROR) {
                     return FRG_STATUS_CLI_ERROR;
                 } else if (result != FRG_STATUS_OK) {
-                    frg_log_prefix_internal();
+                    frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
                     frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to parse next argument: %s", frg_status_to_string(result)); 
                     return result;
                 }
