+++ forge-core/lib/forge/cli/program_demo.c
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License along with Forge.
 // If not, see <https://www.gnu.org/licenses/>.
 
-#include <forge/config/cli/program.h>
+#include <forge/cli/program.h>
 #include <forge/common/log.h>
 
 frg_status_t callback_command(
@@ -44,7 +44,7 @@ int no_commands(void) {
         NULL
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create program: %s", frg_status_to_string(result));
         return 1;
     }
@@ -52,28 +52,28 @@ int no_commands(void) {
     frg_cli_option_t* option = NULL;
     result = frg_cli_option_new_flag(&option, "debug", "Enable debug mode", callback_option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create option: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_option_set_add_option(program->global_options, option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to add option to global option set: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_print_help(program, NULL);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print program help: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_destroy(&program);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy program: %s", frg_status_to_string(result));
         return 1;
     }
@@ -96,7 +96,7 @@ int one_command_global(void) {
         NULL
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create program: %s", frg_status_to_string(result));
         return 1;
     }
@@ -104,14 +104,14 @@ int one_command_global(void) {
     frg_cli_option_t* option = NULL;
     result = frg_cli_option_new_flag(&option, "debug", "Enable debug mode", callback_option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create option: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_option_set_add_option(program->global_options, option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to add option to global option set: %s", frg_status_to_string(result));
         return 1;
     }
@@ -125,7 +125,7 @@ int one_command_global(void) {
         callback_command
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to create command: %s", frg_status_to_string(result));
         return 1;
     }
@@ -133,35 +133,35 @@ int one_command_global(void) {
     option = NULL;
     result = frg_cli_option_new_flag(&option, "option", "An option", callback_option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to create option: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_option_set_add_option(command->option_set, option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to add option to option set: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_add_command(program, command);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to add command to program: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_print_help(program, NULL);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print program help: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_destroy(&program);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy program: %s", frg_status_to_string(result));
         return 1;
     }
@@ -184,7 +184,7 @@ int one_command_command(void) {
         NULL
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create program: %s", frg_status_to_string(result));
         return 1;
     }
@@ -192,14 +192,14 @@ int one_command_command(void) {
     frg_cli_option_t* option = NULL;
     result = frg_cli_option_new_flag(&option, "debug", "Enable debug mode", callback_option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create option: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_option_set_add_option(program->global_options, option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to add option to global option set: %s", frg_status_to_string(result));
         return 1;
     }
@@ -213,7 +213,7 @@ int one_command_command(void) {
         callback_command
     );
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to create command: %s", frg_status_to_string(result));
         return 1;
     }
@@ -221,35 +221,35 @@ int one_command_command(void) {
     option = NULL;
     result = frg_cli_option_new_flag(&option, "option", "An option", callback_option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to create option: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_option_set_add_option(command->option_set, option);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to add option to option set: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_add_command(program, command);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_FATAL_ERROR, "unable to add command to program: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_print_help(program, "command");
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to print program help: %s", frg_status_to_string(result));
         return 1;
     }
 
     result = frg_cli_program_destroy(&program);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to destroy program: %s", frg_status_to_string(result));
         return 1;
     }
