+++ forge-core/lib/forge/codegen/codegen.cpp
@@ -13,19 +13,34 @@
 // You should have received a copy of the GNU General Public License along with Forge.
 // If not, see <https://www.gnu.org/licenses/>.
 
-#pragma once
+#include "generators.hpp"
+#include <iostream>
+#include <llvm/IR/Module.h>
 
-#include <forge/common/enums.h>
-#include <glib.h>
+extern "C" {
+#include <forge/codegen/codegen.h>
 
-typedef struct {
-    bool version_short;
-} frg_config_t;
+frg_status_t frg_codegen(frg_llvm_module_t** module, const frg_ast_t* ast) {
+    return FRG_STATUS_OK;
+}
 
-frg_status_t frg_config_new_default(frg_config_t** config);
-void frg_config_destroy(frg_config_t** config);
+frg_status_t frg_codegen_call_function(
+    void* returned_value,
+    frg_llvm_module_t* module,
+    const char* name,
+    GList* pos_args
+) {
+    return FRG_STATUS_OK;
+}
 
-frg_status_t frg_config_parse_cli(int* exit_status, frg_config_t* config, int argc, const char** argv);
-frg_status_t frg_config_parse_env(frg_config_t* config);
+frg_status_t frg_codegen_write_object_file(
+    frg_llvm_module_t* module,
+    const char* path
+) {
+    return FRG_STATUS_OK;
+}
+}
 
-frg_status_t frg_config_log_debug(const frg_config_t* config);
+struct frg_llvm_module_t {
+    llvm::Module* llvm_module;
+};
