+++ forge-core/lib/forge/codegen/generators.hpp
@@ -15,17 +15,12 @@
 
 #pragma once
 
-#include <forge/common/enums.h>
-#include <glib.h>
+#include <llvm/IR/Type.h>
 
-typedef struct {
-    bool version_short;
-} frg_config_t;
+extern "C" {
+#include <forge/ast/ast.h>
+}
 
-frg_status_t frg_config_new_default(frg_config_t** config);
-void frg_config_destroy(frg_config_t** config);
-
-frg_status_t frg_config_parse_cli(int* exit_status, frg_config_t* config, int argc, const char** argv);
-frg_status_t frg_config_parse_env(frg_config_t* config);
-
-frg_status_t frg_config_log_debug(const frg_config_t* config);
+namespace frg {
+llvm::Type* generate_type(frg_ast_t* ast);
+}
