+++ forge-core/src/forge/main.c
@@ -20,7 +20,7 @@ int main(int argc, char *argv[]) {
     frg_config_t* config = NULL;
     frg_status_t result = frg_config_new_default(&config);
     if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to create config: %s", frg_status_to_string(result));
         return 1;
     }
@@ -30,21 +30,21 @@ int main(int argc, char *argv[]) {
     if (result == FRG_STATUS_CLI_ERROR) {
         return 1;
     } else if (result != FRG_STATUS_OK) {
-        frg_log_prefix_internal();
+        frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to parse config from CLI: %s", frg_status_to_string(result));
         return 1;
     } else if (exit_status != 0) {
         return exit_status;
     }
 
-    if (config->debug) {
-        result = frg_config_log_debug(config);
-        if (result != FRG_STATUS_OK) {
-            frg_log_prefix_internal();
-            frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to log config: %s", frg_status_to_string(result));
-            return 1;
-        }
-    }
+    // if (config->debug) {
+    //     result = frg_config_log_debug(config);
+    //     if (result != FRG_STATUS_OK) {
+    //         frg_log_prefix_internal(FRG_LOG_SEVERITY_INTERNAL_ERROR);
+    //         frg_log(FRG_LOG_SEVERITY_INTERNAL_ERROR, "unable to log config: %s", frg_status_to_string(result));
+    //         return 1;
+    //     }
+    // }
 
     if (!frg_log_summary_if_errors()) {
         return 1;
